UI: mqtt version refactoring.
