+++ ui-ngx/src/app/shared/models/device.models.ts
@@ -902,17 +902,3 @@ export const getAlarmScheduleRangeText = (startsOn: Date | number, endsOn: Date 
   return `<span><span class="nowrap">12:00 AM</span> – <span class="nowrap">${end.format('hh:mm A')}</span>` +
     ` and <span class="nowrap">${start.format('hh:mm A')}</span> – <span class="nowrap">12:00 PM</span></span>`;
 };
-
-export enum MqttVersion {
-  MQTT_3_1 = 'MQTT_3_1',
-  MQTT_3_1_1 = 'MQTT_3_1_1',
-  MQTT_5 = 'MQTT_5'
-}
-
-export const MqttVersions = Object.values(MqttVersion);
-
-export const MqttVersionTranslation = new Map<MqttVersion, string>([
-  [MqttVersion.MQTT_3_1, 'MQTT 3.1'],
-  [MqttVersion.MQTT_3_1_1, 'MQTT 3.1.1'],
-  [MqttVersion.MQTT_5, 'MQTT 5.0']
-]);
